server/src/admin/agents/requirements.txt
crewai
crewai-tools
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary # For PostgreSQL database introspection
